+++ Installation/cmake/modules/CGAL_SetupBoost.cmake
@@ -17,9 +17,10 @@ set ( CGAL_Boost_Setup TRUE )
 
 include(${CMAKE_CURRENT_LIST_DIR}/CGAL_TweakFindBoost.cmake)
 
-find_package( Boost 1.72 REQUIRED )
+cmake_policy(VERSION 3.12...3.30)
+find_package( Boost 1.74 REQUIRED )
 
-if(Boost_FOUND AND Boost_VERSION VERSION_LESS 1.72)
+if(Boost_FOUND AND Boost_VERSION VERSION_LESS 1.74)
   if(DEFINED Boost_DIR AND NOT Boost_DIR)
     # Unset that cache variable that is set in the cache by FindBoost
     # (while it was searching for boost-cmake).